simplify colors
