+++ package.json
@@ -22,9 +22,9 @@
 			"require": "./dist/index.cjs"
 		},
 		"./colors": {
-			"types": "./dist/colors/index.d.ts",
-			"import": "./dist/colors/index.js",
-			"require": "./dist/colors/colors.cjs"
+			"types": "./dist/colors.d.ts",
+			"import": "./dist/colors.js",
+			"require": "./dist/colors.cjs"
 		}
 	},
 	"scripts": {
