--- src/colors/black.ts
@@ -1,16 +1,0 @@
-import type { Color } from '../types';
-import { createColor } from '../utils';
-
-export const BLACK: Color = createColor({
-	hex: 0x000000,
-	rgb: {
-		r: 0,
-		g: 0,
-		b: 0
-	},
-	hsl: {
-		h: 0,
-		s: 0,
-		l: 0
-	}
-});
