--- src/colors/index.ts
@@ -1,5 +1,0 @@
-export { RED } from './red';
-export { GREEN } from './green';
-export { BLUE } from './blue';
-export { WHITE } from './white';
-export { BLACK } from './black';
